Rename meta keys and update tests
Change meta payload keys in make_meta: replace "message" with "title" and "base_url" with "base". Update tests to expect the new "title" key and add a test for /prospects list. Remove the redundant tests/prospects/test_prospects.py file to consolidate test coverage.
